c#移植bitfkags

+++ java-commons/base/src/main/java/cn/wjybxx/base/BitFlags.java
@@ -131,19 +131,19 @@ public class BitFlags {
     }
 
     /** 是否设置了指定下标的bit */
-    public static boolean isSetAt(long flags, long idx) {
+    public static boolean isSetAt(long flags, int idx) {
         long mask = 1L << idx;
         return (flags & mask) != 0;
     }
 
     /** 是否未设置指定下标的bit */
-    public static boolean isNotSetAt(long flags, long idx) {
+    public static boolean isNotSetAt(long flags, int idx) {
         long mask = 1L << idx;
         return (flags & mask) == 0;
     }
 
     /** 设置指定下标的bit */
-    public static long setAt(long flags, long idx, boolean enable) {
+    public static long setAt(long flags, int idx, boolean enable) {
         long mask = 1L << idx;
         if (enable) {
             return (flags | mask);
